update submodule and patches to 2.3.5
after looking through the changelog and discussions.
the 2 dropped patcher were for 6.17 compatibility, which we needed
before 2.3.5 was released.

Issues addressed apart from kernel compatibility that might be of
interest in our usecases:
* https://github.com/openzfs/zfs/issues/17184 (panic in context of
  many hardlinks/inodes being added)
* https://github.com/openzfs/zfs/issues/17761 (issue with sync-writes
  for the non-default zvol_use_blk_mq module parameter setting)

minimally tested on 2 VMs - one running with our 6.17 kernel, one with
our 6.14 kernel - with this version compiled in.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251124170601.2120680-1-s.ivanov@proxmox.com